format code style flake8

+++ bootcamp/feeds/views.py
@@ -1,17 +1,23 @@
-import json
-from django.shortcuts import render, get_object_or_404
-from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseForbidden
-from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.http import (
+    HttpResponse,
+    JsonResponse,
+    HttpResponseForbidden,
+    HttpResponseBadRequest,
+)
 from django.template.loader import render_to_string
 from django.template.context_processors import csrf
+from django.shortcuts import render, get_object_or_404
 from django.contrib.auth.decorators import login_required
-from bootcamp.feeds.models import Feed
-from bootcamp.activities.models import Activity
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+
 from bootcamp.decorators import ajax_required
+from bootcamp.activities.models import Activity
+
+from .models import Feed
 
 FEEDS_NUM_PAGES = 10
 
-# @login_required
+
 def feeds(request):
     all_feeds = Feed.get_feeds()
     paginator = Paginator(all_feeds, FEEDS_NUM_PAGES)
@@ -20,57 +26,67 @@ def feeds(request):
     if feeds:
         from_feed = feeds[0].id
     return render(request, 'feeds/feeds.html', {
-        'feeds': feeds, 
-        'from_feed': from_feed, 
+        'feeds': feeds,
+        'from_feed': from_feed,
         'page': 1,
-        })
+    })
+
 
 def feed(request, pk):
     feed = get_object_or_404(Feed, pk=pk)
     return render(request, 'feeds/feed.html', {'feed': feed})
 
-# @login_required
+
 @ajax_required
 def load(request):
-    from_feed = request.GET.get('from_feed')
     page = request.GET.get('page')
+    from_feed = request.GET.get('from_feed')
     feed_source = request.GET.get('feed_source')
+    csrf_token = str(csrf(request)['csrf_token'])
+
     all_feeds = Feed.get_feeds(from_feed)
+
     if feed_source != 'all':
         all_feeds = all_feeds.filter(user__id=feed_source)
+
     paginator = Paginator(all_feeds, FEEDS_NUM_PAGES)
+
     try:
         feeds = paginator.page(page)
     except PageNotAnInteger:
         return HttpResponseBadRequest()
     except EmptyPage:
         feeds = []
+
     html = ''
-    csrf_token = str(csrf(request)['csrf_token'])
     for feed in feeds:
-        html = '{0}{1}'.format(html, render_to_string('feeds/partial_feed.html', {
+        context = {
             'feed': feed,
             'user': request.user,
             'csrf_token': csrf_token
-            })
-        )
+        }
+        template = render_to_string('feeds/partial_feed.html', context)
+
+        html = '{0}{1}'.format(html, template)
+
     return HttpResponse(html)
+
 
 def _html_feeds(last_feed, user, csrf_token, feed_source='all'):
     feeds = Feed.get_feeds_after(last_feed)
+
     if feed_source != 'all':
         feeds = feeds.filter(user__id=feed_source)
+
     html = ''
+
     for feed in feeds:
-        html = '{0}{1}'.format(html, render_to_string('feeds/partial_feed.html', {
-            'feed': feed,
-            'user': user,
-            'csrf_token': csrf_token
-            })
-        )
+        context = {'feed': feed, 'user': user, 'csrf_token': csrf_token}
+        template = render_to_string('feeds/partial_feed.html', context)
+        html = '{0}{1}'.format(html, template)
     return html
 
-# @login_required
+
 @ajax_required
 def load_new(request):
     last_feed = request.GET.get('last_feed')
@@ -79,48 +95,60 @@ def load_new(request):
     html = _html_feeds(last_feed, user, csrf_token)
     return HttpResponse(html)
 
-# @login_required
+
 @ajax_required
 def check(request):
     last_feed = request.GET.get('last_feed')
     feed_source = request.GET.get('feed_source')
     feeds = Feed.get_feeds_after(last_feed)
+
     if feed_source != 'all':
         feeds = feeds.filter(user__id=feed_source)
+
     count = feeds.count()
     return HttpResponse(count)
+
 
 @login_required
 @ajax_required
 def post(request):
     last_feed = request.POST.get('last_feed')
+    post = request.POST['post'].strip()[:255]
     user = request.user
+
     csrf_token = str(csrf(request)['csrf_token'])
-    feed = Feed()
-    feed.user = user
-    post = request.POST['post']
-    post = post.strip()
+
     if len(post) > 0:
-        feed.post = post[:255]
-        feed.save()
+        Feed.objects.create(
+            post=post,
+            user=user
+        )
     html = _html_feeds(last_feed, user, csrf_token)
     return HttpResponse(html)
+
 
 @login_required
 @ajax_required
 def like(request):
-    feed_id = request.POST['feed']
-    feed = Feed.objects.get(pk=feed_id)
     user = request.user
-    like = Activity.objects.filter(activity_type=Activity.LIKE, feed=feed_id, user=user)
+    feed_id = request.POST['feed']
+
+    feed = Feed.objects.get(pk=feed_id)
+    like = Activity.objects.filter(activity_type=Activity.LIKE, feed=feed_id,
+                                   user=user)
     if like:
         user.profile.unotify_liked(feed)
         like.delete()
     else:
-        like = Activity(activity_type=Activity.LIKE, feed=feed_id, user=user)
-        like.save()
+        Activity.objects.create(
+            feed=feed_id,
+            user=user,
+            activity_type=Activity.LIKE
+        )
         user.profile.notify_liked(feed)
+
     return HttpResponse(feed.calculate_likes())
+
 
 @login_required
 @ajax_required
@@ -128,19 +156,22 @@ def comment(request):
     if request.method == 'POST':
         feed_id = request.POST['feed']
         feed = Feed.objects.get(pk=feed_id)
-        post = request.POST['post']
-        post = post.strip()
+        post = request.POST['post'].strip()
+
         if len(post) > 0:
             post = post[:255]
             user = request.user
             feed.comment(user=user, post=post)
             user.profile.notify_commented(feed)
             user.profile.notify_also_commented(feed)
-        return render(request, 'feeds/partial_feed_comments.html', {'feed': feed})
-    else:
-        feed_id = request.GET.get('feed')
-        feed = Feed.objects.get(pk=feed_id)
-        return render(request, 'feeds/partial_feed_comments.html', {'feed': feed})
+
+        context = {'feed': feed}
+        return render(request, 'feeds/partial_feed_comments.html', context)
+
+    feed_id = request.GET.get('feed')
+    feed = Feed.objects.get(pk=feed_id)
+    return render(request, 'feeds/partial_feed_comments.html', {'feed': feed})
+
 
 @login_required
 @ajax_required
@@ -148,14 +179,19 @@ def update(request):
     first_feed = request.GET.get('first_feed')
     last_feed = request.GET.get('last_feed')
     feed_source = request.GET.get('feed_source')
+
     feeds = Feed.get_feeds().filter(id__range=(last_feed, first_feed))
+
     if feed_source != 'all':
         feeds = feeds.filter(user__id=feed_source)
+
     dump = {}
+
     for feed in feeds:
         dump[feed.pk] = {'likes': feed.likes, 'comments': feed.comments}
-    data = json.dumps(dump)
-    return HttpResponse(data, content_type='application/json')
+
+    return JsonResponse(dump, safe=False)
+
 
 @login_required
 @ajax_required
@@ -164,23 +200,28 @@ def track_comments(request):
     feed = Feed.objects.get(pk=feed_id)
     return render(request, 'feeds/partial_feed_comments.html', {'feed': feed})
 
+
 @login_required
 @ajax_required
 def remove(request):
-    try:
-        feed_id = request.POST.get('feed')
-        feed = Feed.objects.get(pk=feed_id)
+    feed_id = request.POST.get('feed')
 
-        if feed.user == request.user or request.user.is_superuser:
-            likes = feed.get_likes()
-            parent = feed.parent
-            for like in likes:
-                like.delete()
-            feed.delete()
-            if parent:
-                parent.calculate_comments()
-            return HttpResponse()
-        else:
-            return HttpResponseForbidden()
-    except Exception as e:
-        return HttpResponseBadRequest()
+    feed = Feed.objects.filter(pk=feed_id).first()
+
+    if not feed:
+        return HttpResponseBadRequest
+
+    if feed.user == request.user or request.user.is_superuser:
+        likes = feed.get_likes()
+        parent = feed.parent
+
+        for like in likes:
+            like.delete()
+
+        feed.delete()
+        if parent:
+            parent.calculate_comments()
+
+        return HttpResponse()
+
+    return HttpResponseForbidden()
